backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-celery-results>=2.6.0",
    "django-environ>=0.13.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-guid>=2.2.1",
    "django-import-export>=4.4.0",
    "django-postgres-extra>=2.0.4",
    "django-structlog>=10.0.0",
    "django-unfold>=0.85.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-api-logger>=1.1.21",
    "drf-extensions>=0.7.1",
    "drf-spectacular>=0.29.0",
    "drf-writable-nested>=0.7.2",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.3",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "nplusone>=1.0.0",
    "slack-sdk>=3.41.0",
    "uvicorn[standard]>=0.42.0",
    "channels>=4.2",
    "channels-redis>=4.2",
    "django-crequest>=2018.5.11",
    "boto3>=1.38.0",
    "django-storages[s3]>=1.14.0",
    "django-allow-cidr>=0.8.0",
    "django-cors-headers>=4.9.0",
    "django-markdownx>=4.0.9",
    "django-easymde>=1.0.4",
    "pgvector>=0.4.2",
    "openai>=2.30.0",
    "langchain>=1.2.15",
    "langchain-core>=1.2.27",
    "langchain-community>=0.4.1",
    "langchain-aws>=1.4.3",
    "langchain-openai>=1.1.12",
    "hypothesis>=6.151.9",
    "faiss-cpu>=1.13.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.3",
    "pycurl>=7.45",
    "pydot>=4.0.1",
]

[dependency-groups]
test = [
    "coverage>=7.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.151.9",
]
